packages/backend/pyproject.toml
[tool.poetry]
name = "form-sales-support-backend"
version = "0.1.0"
description = "フォーム営業支援システム バックエンドAPI"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# Webフレームワーク
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}

# データベース
sqlalchemy = "^2.0.35"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}

# バリデーション
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"

# 認証
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

# ユーティリティ
python-multipart = "^0.0.20"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
# テスト
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
testcontainers = {extras = ["postgresql"], version = "^4.9.0"}
httpx = "^0.28.0"

# リンター・フォーマッター
ruff = "^0.8.0"
black = "^24.10.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
